Moved tests to a test-folder and changes to 'for'
