+++ tests/DockFilterTest.js
